# chore: add benchmark showing a 2.5 reduction in memory usage since RC2