@withinboredom 2023-11-10
> A possible way to improve this could be to populate the CGI variables directly in C-compatible strings

If you add a `"\0"` onto the end of the go strings (strings in Go are not null terminated), and know the strings won't be mutated ... you could probably try this super hacky (untested) method to get the raw pointer from Go:

```
char* cstring = (*reflect.StringHeader)(unsafe.Pointer(&gostring)).Data
```

Maybe?